Hid Nugi-chan's newest collectible
Nugi-chan's newest collectible requires beating her in a seasonal costume, but we save these costumes for events to help them be more special. As this would make the collectible unobtainable on spnati.net, I've made it hidden for now.

Alternatively, I could have given it a secondary way to unlock without the skin. But this was the quickest way to help avoid collectible-hunter consternation.