Moon's epilogue, now in HD!
"Surrounded on the Streets" (male version)
- original DK scenario model posing recreated by Phonebox with the glorious new Moon model
- HD conversion and typesetting by Arndress
- support for the Full Moon skin color
- no longer stops at the part where she invites sex!

The female version of this epilogue with follow in a future update.

Moon is still on testing, and so is this epilogue. Give it a try and let us know what you think!

(A number of lines of dialogue appear to have shifted positions with this update. This is because I removed manual priority from these where it was not needed. No lines were removed.)