Epilogues: add if-gallery directive condition
Extends the condition groups for directives to add <if-gallery>
conditions.

These test if the epilogue is currently being played from the
Gallery or not: a value of "true" indicates that the condition
matches when playing the epilogue from the gallery; all other values
match when playing the epilogue after a game.