Made Moon less complainy: Hand lines
A good piece of feedback from episode 42.5.5 of the SPNatI podcast is that Moon complains too much when she's playing, and that's making her not as fun to play against as she could be.

I am going to be doing adjustments to Moon that limit how much she complains by spitting her lines into two groups,  more complainy and less complainy, and limiting the more complainy lines for times when she's stuck in a stage for too many rounds. This will retain the diversity of lines she has in these cases to reduce repetition when on a stage for longer than usual, but it will also help her seem more aware of how long she's had to wait between turns.

Moon's final stage dialogue is overall supportive and not so hungry for the spotlight, but it's the journey there that is needing refinement.

This commit only splits the cases of her Good/Okay/Bad Hand lines. Despite a large amount of shuffling in the diff, no lines were added or removed.