Made Moon less complainy: Must Strip (Self) lines
See the previous Moon update note for more details.

The way Moon's Must Strip (Self) lines were written presumed that she'd been waiting if she was Winning, waiting less in Normal circumstances, and ahead of the pack in Losing cases. This translated perfectly to combining all of these into the general Must Strip (Self) case and allowing her to play the impatient lines only if she gets stuck on a particular stage for several rounds. This will achieve the overall goal of getting her to be less hostile, especially when it's unwarranted.

No lines actually added or removed in this update, just conditional shuffling again.