Re-use updateIndividualSelectVisibility() from updateIndividualSelectFilters() and fix autoclear.
selectableOpponents is no longer used.

There's no need to check if there's any filter values to clear before calling clearSearch(); there can't be any infinite recursion.

Only thing that's suboptimal is that we go on and hide all selection cards before discovering that there are no ones left visible, so we have to clear the filter.