Show the separator only if there are selection cards both above and below it
by sorting loadedOpponents (we can do this now since we don't allow the user to enter the selection screen before all opponent metadata are loaded) and keeping track of before which index the separator is.

If separator is in use, we have to call updateIndividualSelectVisibility() after changing the sort order so it can be hidden when needed.