Sorting loadedOpponents in place broke Featured sort order. Fix it by storing the indices within the Opponent objects.
Also remove a level of indirection by putting the sort keys in the data-value attribues of the dropdown menu items.

Also change the alphabetical sort orders to sort by label, because people have found sorting by the less visible full names to be confusing.

This changes sortOpponentsByMultipleFields to expect an array. On the other hand, there's no reason to use `sortOpponentsByMultipleFields` to sort by a single field.