Daria: Merged 46 duplicated hand cases
The CE has a useful tool to merge targeted lines that are in all three of the Good/Okay/Bad Hand cases into the newer Any Hand case. The CE's Dashboard offers to do this if the name you have in the settings matches the name of the character's author/artist in the metadata.

This won't cause a change to Daria's gameplay, but it does mean that any corrections made to her lines now don't have to be made in three places.