Added the masquerade_candidate tag to 41 characters
The Rework Masquerade is coming sometime early next year, and the candidate list is growing steadily. To give you a better visualization than reading throught the contents and comments of a Reddit thread, I've added the `masquerade_candidate` tag to the 41 thus far confirmed to be participating. Now you can search for this tag in the full roster of your offline version and see images of all of the characters at a glance. And this list will surely grow before applications open.

I will remove this temporary tag when the event begins. These characters are not necessarily abandonned or being put up for adoption for whomever whenever. For some of them, the event is a unique opportunity to take on their development. Any characters that are not assigned to teams during the event remain under the control of those who submitted them.

As a reminder: the Felicia we have in the game right now does not have 3+ layers, so she would not be able to be sponsored to the main roster. This doesn't mean that she couldn't be improved, however.