KKL: Update v104.1 zipfile
Updates KKL to finalize / fix the `direct-screenshot` command,
and also adds documentation for it.

The `direct-screenshot` command is a new method for taking screenshots
via the remote control protocol that is better optimized and also
has more options.

In particular, the `direct-screenshot` command accepts an option
enabling the use of a faster, but less space-efficient PNG encoder;
this drops the time required to take a screenshot down to about
two-tenths of a second (from over two seconds), at the cost of
making output images about 1.4x to 1.6x larger.

However, the larger image sizes probably won't be noticeable if you're
compressing your images with external tools (as you should be), or
if you're doing image manipulations (which would require you to
re-encode the final image anyways).

In conjunction with the `fastload` command, this should help
significantly cut down on the time required to generate character
images in many cases.