Avoid problems with insufficient prefill suggestions
If, for example, there are only three characters eligible to be prefilled, when we got to the last slot, loadDefaultFillSuggestions() would be called again, jumbling the already assigned prefills. That was only supposed to happen when no prefills had been made yet.

Now we avoid stealing prefills from other slots, clear prefills when a character is selected or a prefill is selected in another slot, and only pick new prefills when one is not already picked.

As a side effect, removing a selected character lets you get a new prefill suggestion.