Add variable to help with correct indefinite articles
This also allows clothing items to have `plural="null"` for uncountable nouns like "armor" and "jewelry".

`~clothing.withart~` expands to the same as `~clothing~` but with "a" or "an" in front if singular and countable. Examples: "an eyepiece", "a uniform", "pants", "armor".

The existing `~clothing.toplural~` is extended not to do anything with uncountable nouns.

The existing `~clothing.plural~` is extended to return "uncountable" for uncountable items.

Existing `~clothing.ifplural()~` is extended with an optional third argument that is used instead of the second argument if it exists and the clothing is uncountable.

(Also adds correct pluralization (I hope) for words ending in -y, although no such clothing exists in the game at this point.)