Updated Panty's tags
Also added a new batch of 10 males for Panty to recognize for her book collectible, but no matter how many times I tested this, I could *not* get it to work. The collectible definitely worked when it launched. Does it work for you? Please let me know in Panty's channel.