N Horny Boy Update [From Thursdays]
n is now more horny ladies and gentlemen. also he has a collectible.

When both N and the player are masturbating, he will show more attraction to the player. When N is in the heavy masturbation stage, N will have lewd fantasies and/or dirty-talk the player, but only under the following circumstances:

They have 4 or less layers on.

They have the gay or bisexual tag (only applies if you play as a male)