Adds Ruka as an Offline-Only Opponent
Adds Ruka, a male character from Steins;Gate, as an "incomplete" character by Stuart and Walter.

Ruka meets all the requirements for Testing, however his devs do not have the time or interest in fleshing him out for the main game. However, he is available for adoption by anyone interested.

If you are interested in giving this twink what he needs, just need to lay it on this foundation!

- Also, moves Mordred offline and secretly fixes a typo in one of the docs.