Nagisa costume upgrade!
Nagisa's Valentine's Day has now met the modern model with a beautiful new outfit from HorseKiller! This is based on Nagisa's canon waitress uniform in Clannad. The new outfit is the work of HorseKiller, and it was adapted to her existing poses by yours truly... with some more consultation by HK! The Pose Matrix makes this a less painful process.

I also want to give a shoutout to UnderscorM3, whose adorable and winsome outfit Nagisa previously wore for years. UnderscorM3 is a master of his craft. I also want to acknowledge Patisdom, who previously did many of the stages of fitting her original costume to layers 2+ in a pre-Pose Matrix world.