Saved several ancient characters in CE v6.5
Most of these characters were last saved in some of the version earliest versions of the Character Editor, usually in 2018.

From spnati_edit:
> the behaviour.xml format change introduced that has completely hamstrung me from being able to add a bunch of things that have been on my todo list for a long time.

Because these ancient files aren't in the right format, it makes it very difficult for spnati_edit to cater to every different format at once. The modern format is vastly superior to the old format. This is a step in the direction of standardizing them.