Moved Gwen and Marceline to the Testing Roster for the Rework Masquerade!
Gwen is being reworked by the "Eiffel 85" team and is the seventh character from the Rework Masquerade event to enter the Testing Roster.

Marceline is being reworked by the "Hambo's Himbos" team and is the eighth character from the Rework Masquerade event to enter the Testing Roster.

Two more to go!