Moved amy_rose to amy_rose_classic
Moving a folder name to the `_classic` suffix is not something we do for replaced characters much anymore because it can break so many existing conversations. However, in the case of Amy Rose, who has a new version coming, the old Amy Rose had exactly *zero* outbound targeted lines. Additionally, many of the inbound targeted lines can be used by the new Amy Rose, even those from characters with long-dead creators.

However, a small number of these inbound lines are specific to the original version of Amy Rose and won't apply to the model. Kobrad has identified 12 characters who either mention something specific about the old Amy's clothing or who target her clothing in a stage that no longer applies. For 9 of these characters, fixes are in this commit to make sure they correctly address these to amy_rose_classic. For Jessie, Pit, and Palutena, we leave this up to their active authors to fix.