Fixes for Blake
It's my understanding that Blake has an update in eternal progress, so this is not an exhaustive fix for Blake, but it does pluralize her wardrobe and put a Play Once on most of her lines to Kali in cases where it's possible to repeat.

If an update does happen and it uses the same folder, it can safely overwrite this commit.

Note: The author for Blake doesn't seem to have been fully familiar with how SPNatI dialogue works, so sometimes Blake appears to be responding to specific lines from Kali without any way to check if they're actually being said. This update makes to attempt to untangle this.