Kali fixes
This update endeavors to free up more dialogue for Kali that either couldn't play or could only play in unintentionally-strict circumstances.

Additionally, I did a little bit more to simplify the conditions on some of her dialogue. For example, if she's checking for exactly 1 female (herself), she doesn't need to also to check for 2-4 males, as that's inevitable because SPNatI is a 3-5 player game.

Her line count has fluctuated up and down a few lines with these fixes. The likely cause is that sometimes when an error is corrected, the new form matches a line where that error had already been corrected in a previous update. And the reverse is true: sometimes an error is only fixed in one case, making the unique line count grow misleadingly.