Saved more characters in CE v6.5.1
In addition to the previous commit, the following characters have now been saved in CE v6.5.1. This brings all characters in line with the internal sort-by-case structure for behaviour.xml files first applied by CE v5.1.

* Calisto: CE v4.4b -> CE v6.5.1
* Chiaki: CE v5.0 -> CE v6.5.1
* Chica: make_xml v1.55 -> CE v6.5.1
* Corrin M: CE v3.4 -> CE v6.5.1
* Kamina: CE v5.0b -> CE v6.5.1
* Kyoko: CE v5.0 -> CE v6.5.1
* Meg: CE v5.0 -> CE v6.5.1
* Misaka: CE v5.0 -> CE v6.5.1
* Navi: CE v4.0 -> CE v6.5.1
* Rainbow Mika: CE v5.0 -> CE v6.5.1
* Rinkah: CE v5.0 -> CE v6.5.1

I also took this opportunity to remove make_xml.py and make_images.py. Though both were once essential for creating characters in the years before the Character Editor, the latter has not been used in a long time. The former is no longer the basis for any character files as of this commit! It's been a long road getting to this point, but we made it!