Promotionally returns Monika to the Main Roster
DDLC received an update today that we anticipate will see increased interest in the DDLC girls, including Monika. As she was previously on the Main Roster before her return to the Testing Roster, Monika has temporarily relocated herself to the Main Roster to benefit from the attention for the next two weeks.