Use custom poses from default costume in stages an alternate costume specifies that the base folder be used instead of the costume folder.
This moves assignment `.poses` to `.updateFolder()`. To avoid crashes, and because it's unnecessary, don't call `.stageChangeUpdate()`, which calls `.updateFolder()`, until behaviour has been loaded.

When we're at it, don't call `.resetState()` at all from the constructor, as it's not needed until a character is loaded. This does mean that we have to call it on `humanPlayer` explicitly, though.