Prevent rolling back in the middle of dealing cards, AI turns, etc., because that messes up everything.
This is done by changing what `actualMainButtonState` means: It's now `true` when we're in the middle of those things, but not while waiting for the auto forfeit timeout (the variable could get a better name).

Also pauses auto-advance whenever the bug report, feedback, restart, or game log modals are open, as well as when rolled back. This simplifies the handling when closing the Options modal.

One complication is when clicking Feedback in the Restart modal and then going back, thereby hiding one modal and showing another at the same time. It's then important to cancel the auto forfeit timeout again *after* setting it, which is accomplished by using the `hide.bs.modal` and `shown.bs.modal` handlers.