Add a new "unsorted-updated" highlight
"unsorted-updated" is to "updated" as "unsorted" is to "new" - that is, it keeps them auto-suggested (to compensate for being at the bottom of the roster), but without the big "Major Update" flair.

Has to be separate from just Unsorted so that it displays "Updated" on the selection card.

First use will be for WFT, for the few days between her highlight expiration and the re-sort.