Samus: Added markers and corrected timestamp
Sort by Last Updated in offline mode to see when Samus last had a non-maintenance update.

Samus has been enduringly popular despite a lack of recent updates, and this makes her difficult to interact with. Samus now has markers on her targeted lines to help you latch on to this a little easier, though most of the characters she targets are also inactive