Ryu: Compressed pose images
Also removed the black text from the background of one of his epilogue images; this is dialogue he already says in this scene and appears to have been left in by accident. It's difficult to see, but we had a bug report about it.

While I was here, I corrected timestamp to Ryu's last non-maintenance update.

No tag changes; these re-sorted automatically.