Events: fold Event sorting into default Featured sorting
This removes the "Event" sorting mode, and folds its functionality
into the "Featured" sorting mode (for both the main and testing rosters)
and into the "Last Updated" sorting mode (for the testing roster only).

This makes things a bit more seamless, in case events are used for
things like new character highlighting; since these aren't _major_
events, it might be a bit confusing to show an "Event" mode
in the sort options dropdown box.