Lyralei: added arcana-model scene, updated pubic hair, various bug fixes
Added a scene where Lyralei shows her Arcana in a normal game instead of just the epilogue every third game if the player wins that game. Also updated all of hair her pubic her with some fixes to other images and lines.

Lyralei's file space currently over the 11mb limit (by about 700kb), which is mostly due to images that are/will be used in the epilogue as well (her arcana pose, the wind cloud for the animation, etc.). - let me know if I need to remove some other poses to get back under the 11mb.