Add ~blank~ variable that expands to nothing
Can be used to create a blank line; currently, a line consisting only of "~silent~" in the CE results in a blank line being saved, as does a blank line as long as you choose a pose. ~blank~ lets you indicate an intentionally blank line, so the CE's validator can check for unintentionally blank lines.