Mari: add memory
Adds special Game Start lines that vary based on the last
time the player started a game against Mari, as well as
the winner of the last game (if any).

Specific situations targeted are:
- Player won last game
- Mari won last game
- Other character won last game
- Last game was unfinished

Each situation has two variants, depending on if the player's
having a rematch (on the same day as the last game), or is otherwise
playing a match on a different day as the last.

There are also lines for players who play 3, 4, 5, or 6+ matches
against Mari in a day.