Fix variable tests with empty RHS values
Changes variable tests to use an empty string as the RHS, instead
of `null`, when no `value` attribute is present.

Strictly speaking, such tests should use `~blank~` as the RHS value,
but I don't think there are any characters that depend on the previous
behavior in this case.

Also changes Mari's cases testing against `~target.wearing(shoes)~`
to test against `~blank~` as the RHS, instead of an empty RHS.