Add counter for player ticks in stage
This counter is incremented (for each player) whenever
tickForfeitTimers() is called, and is set to 0 just before a player's
Stripped or Start Masturbating dialogue is played.

In general, when is player is masturbating,
their ticksInStage + their timer = their stamina value.

This includes when a player is finishing, at which point ticksInStage
should equal their stamina value.

Like the timeInStage counter, this is only reset to 0 just after
the Finished cases are played.