Mari: adjust dynamic difficulty settings
Adjusts Mari's hidden cases for setting her difficulty with 2+
players left to work as follows:

- If she's ever 3 or more layers behind the best-performing player,
regardless of stage, her AI will be set to Best.

- In her late stages (stages 5-7), her AI will follow its normal
difficulty course if she's less than three layers behind the lead
player; this means that she'll be Good in stages 5 and 6, and then
will always be Best in stage 7.

- In her early stages (stages 0-4), her AI will be reset to Average
if she's only 1 layer behind or in first place.

- However, if she's exactly 2 layers behind in her early stages,
her AI will be set to Good, as an intermediate step.