Halloween 2021: Elevated characters with new costumes to the top of the roster
The Event Schedule seems to have worked well, but an issue that wasn't obvious until a couple of days ago is that there doesn't seem to be way to super-promote characters who have new costumes for 2021. So this is an old-fashioned manual adjustment of listing.xml to do just that.