Clothing: begin work on extended clothing options
Adds a PlayerClothing class, which is a subclass of Clothing.

The title screen has been reworked so that it is dynamically
populated at runtime based on a list of PlayerClothing objects.

The code for saving and loading player wardrobes has also been
reworked to provide a better interface when working with PlayerClothing.