Clothing: add `~clothing.id~` variable
Adds `~clothing.id~`, which will return the ID of the clothing
a human player is removing / has removed. `~revealed.id~` works
similarly.

For default clothing, this will expand to an ID beginning with
`_default`, followed by a dot and then a sub-ID (for example,
`_default.boots`).

For unlockable clothing, this will expand to the ID of the associated
collectible.

In cases where the targeted clothing has no ID (i.e. when targeting
AI players), this will expand to an empty string.