Collectibles: add ~collectible.*.wearing~ variable
This variable expands to `true` if the human player is currently wearing
the clothing item associated with the given collectible, and expands
to `false` otherwise.

For collectibles without associated clothing items, this always expands
to `false`.

This can be used to test if the player's wearing an unlockable piece
of clothing before it is removed or revealed.