Switch new font to Rubik instead of Noto Sans
This commit actually adds a variable font version of Rubik, instead
of the usual static weight versions. This means that we can use a much
wider range of font weights as opposed to the standard 400 ("regular")
and 700 ("bold") weights.

To be specific, the supported range of weights for Rubik goes from
300 ("light") to 900 ("black"). It'll be interesting to see how we
can take advantage of this in the future.

(All modern browsers should support variable fonts; however, in the
unlikely case where a player can't use variable fonts, they'll fallback
to Trebuchet MS instead.)