Fonts: switch to Open Sans
Switches to Open Sans from Rubik.

Open Sans is another variable font, but it also supports changing
text width in addition to font weight. This commit adds sliders for
allowing the user to adjust default values for both of these values,
in the in-game options menu.

This commit retains the old definition and files for Rubik, just
in case we decide to switch back.