Don't consider opponent_lost in parallel with female_human_must_strip and male_human_must strip.
It doesn't make sense to consider opponent_lost twice, because this way, (fe)male_must_strip would never be considered when the human player has to strip if there are matching opponent_lost cases.

With this change, (fe)male_must_strip will be considered in parallel with opponent_lost, just like when a character has to strip.