Instead of stopping after finding a matching (fe)male_human_must_strip case, consider them together with (fe)male_must_strip and opponent_lost as if they had a target="human" condition
This means giving such cases +300 to their default priority.

This in turn requires the `Case` constructor to get the trigger ID as an argument.

We avoid adding 300 to the priority if the case already has a `target: human` condition (old or new style), but we don't actually _convert_ the case to `[fe]male_must_strip`.

Also rename another variable and a property from `tag` to `trigger`.