Derive game "last updated" time from CI/CD
Changes the "last updated" time shown in the title screen (below
the game version number) to display timestamps that are automatically
filled in by the CI/CD pipeline.

This doesn't involve adding any versions or changing details of existing
versions in the log; this only affects the update time shown on the
title screen.

Offline versions of the game will still pull their "last updated"
timestamps from the latest version in the update log, as before.

This should hopefully make the actual update pace of the game much
clearer to casual players / non-devs.