Jin: Added an image for the Goggles collectible
Zeuses Swan Song sent this image to me in October 2021, and I just now got around to attaching it. I also made this collectible wearable at the same time. However, this is a secret collectible, so it won't show up in the player wardrobe list unless you unlock it first; it's just a nice perk for people who were going to play against Jin anyway.

The reason why it's a secret collectible is because it relies on knowing when Jin sees a butt. This was previously handled by a lot of targeted lines, and this quickly gets out of date as new characters join the game. I've made a small improvement that will unlock this collectible if Jin sees any character with an exposed crotch using a pose named "butt", "rear", "ass", "booty", or "back". I made the TCNatI Butt card set, so I know these are the butt pose names used by more than one character. Other edge cases may arise in the future, so I'll leave the collectible hidden until unlocked.