Mari: add lines for consecutive self losses
Adds Must Strip (Self) lines for when Mari loses multiple rounds
in a row. Every possible number of rounds lost in a row is covered
(2 through 7).

Situations where Mari has lost two or three rounds in a row are
covered by two different cases (each), depending on Mari's current
stage--the extra variation is handy because of the relative frequency
of this situation.

Each of the 8 added Must Strip (Self) cases has three lines, and
each of them has Play Once set; none of these lines should repeat
in a given game.

Two Must Masturbate (Self) cases have been added to also cover
consecutive losses; for this situation, I didn't cover the case of
two losses in a row, but I did cover every other possible case,
with two lines each for 3-5 losses in a row and 6+ losses in a row.

In total, this makes for 28 new lines covering situations where
Mari has lost multiple rounds in a row, with full coverage of
every possible number of lost rounds in a row.