Add finishing dialogue redirection
Adds a new forfeit state operation for redirecting a character's
finishing dialogue to another character, accessible by setting
`attr="redirect-finish"` on a `<forfeit>` operation element.

This operation type interprets the `value` attribute of the operation
as a reference to a player (i.e. `self`, `target`, `winner`, a variable
binding, or a character ID), and marks that character as being the
'finishing target' for the character executing the operation.

When the character whose dialogue has been redirected finishes, instead
of playing `finishing_masturbating` dialogue for that character, we
instead look for and play `opponent_finishing_masturbating` dialogue
on the finishing target (targeted towards the player that is actually
finishing, naturally).

This allows joint forfeits to work seamlessly, by allowing characters
that are visually not in their actual table slot to display dialogue
in the correct position.