Add "Finishing (Opponent)" to CE case type list
This particular case is used with the recently-added finishing
dialogue redirection feature (!5558) to allow finishing dialogue
to work correctly during joint forfeits.

This case type isn't actually going to be usable until the CE
adds support for the redirection feature itself, but we might
as well add it now in preparation.