Display number of commits since version number update
Adds a count of commits since the latest tagged version to the
version number displayed online, derived from the internal version tag
number contained in `config.xml` (which itself is filled in during CI/CD
with `git describe`).

For example, as of this commit, this shows up on my machine as
`v12.137.0-9`, indicating that 9 commits have been added to `master`
since we tagged version `v12.137.0`.

If we find a valid, non-zero commit count, we also add a special
entry to the top of the in-game version log describing what that number
actually means in general terms (i.e. that it represents smaller updates
that haven't been specifically noted in the version log yet).

This will hopefully reduce the number of people we get asking why the
'last updated' timestamp keeps changing without a version log update.