CSS: convert dialogue and name label colors to CSS variables
Adds CSS variables for certain colors used in the main game and
selection screens; specifically, the colors used for dialogue bubbles,
name labels, and a few other related labels.

This lets us easily change the color scheme for these elements by
setting these variables to different values; the `dark-mode` CSS class
demonstrates this nicely.

In the future, we may want to consider extending this concept to the
rest of the UI, even if just to help ensure we have a unified color
scheme across the entire game. That'll be a rather significant
undertaking, however.