Add alternate epilogue unlock requirements
Adds a new `unlock` epilogue condition attribute to control the
'baseline' requirement for unlocking an epilogue.

Possible values are, from most to least stringent:
- `player-win`: The human player must win the game. This is equivalent
  to the current default requirement.
- `outlast`: The human player must either win the game, or at least only
  lose *after* the epilogue character.
- `character-loss`: The epilogue character must lose the game.
- `completion`: The game must have been completed.

These correspond to the options discussed in #118.

This does *not* change the default unlock requirement yet; the `unlock`
attribute, if not set, is treated as `player-win`.