May and Cynthia become friends: part 2
There's still a little more polish that could be added, but this is the majority of the new May-Cynthia exchange! Play them as Cynthia-May or May-Cynthia to uncover entirely new conversations!

All lines play when they should, but it's possible they also play when they shouldn't. Please let me know if you experience any bugs.

After a little initial beta testing, Alphaman or I will announce this on the subreddit. Thanks goes to Alphaman for thinking of many interesting and fun things for Cynthia to say!