Nagisa: Removed epilogues
Nagisa's epilogues served her well for many years until her amazing model update. Though it is possible to modernize an old epilogue with a new model, as I did painstakingly and faithfully for Moon, I do not believe that Nagisa's existing epilogues have enough substance to warrant this kind of effort for improvement. They'll live on in our hearts and various outdated rehosted versions of the game.

This update also adds a few lines to Jim Raynor, but these are not yet ready for wider testing, especially as Mr Raynor's half of the conversation is not yet implemented.