Convert old-style Case attributes to Conditions
The Case constructor now automatically converts condition
attributes on the <case> element into Condition objects,
grouping them together where possible (e.g. for the target*
and alsoPlaying* attributes, as well as for saidMarker,
notSaidMarker, and hasHand attributes, which look at self).

As part of this, the Condition constructor has been changed
to take an Object mapping properties to values, instead of
an XML object. XML attribute reading is done in separate
functions now, typically via `conditionFromAttrs`.

This significantly simplifies the Case-level condition checking
code (which will be necessary for some later work).

==

The new post-conversion Condition objects should have the
same priority value and behavior as the old condition
attributes, except in two edge cases:
- `hasHand` for Conditions has a different priority value with
  `role="other"` compared to `alsoPlayingHand`: the Condition
  gives +15 priority, while the Case attribute gives +5.
- Case-level `timeInStage` and `targetTimeInStage` specifically
  check if the checked player's `timeInStage` equals -1, and
  replaces it with 0 if so. This check is omitted when evaluating
  Conditions.

I'm not entirely sure what to do about these discrepancies, so
I'll leave it to a future commit.

==

The old Case-level `consecutiveLosses` condition has unique
behavior compared to its Condition counterpart: for situations
with a target, it will check the target's consecutive loss count,
but if there is no target, it will check the speaker's losses instead.

Since there's no easy way to map this into a preexisting Condition
role, I added a new one: `focus`. This role corresponds to the
target if there is one, or to the speaker otherwise.

For Must Strip/Stripping/Stripped situations, this effectively means
that the role always checks the player that is currently stripping,
regardless of whether that's the speaker or not. In other words,
it checks the "focus" of the current game situation.

This role technically works with all Condition attributes, and can
be used by setting `role="focus"` in XML, though I can't immediately
see any explicit use for this in dialogue.

==

A few Case-level attributes remain:
 - `stage`, since it's used for pre-filtering cases in `findBehaviour`
   (among other things)
 - `totalRounds`, since that condition isn't character-based