spniDisplay: attempt to fix custom pose flickering on Firefox
TL;DR: delay clearing out old poses until the repaint *after* a new
custom pose is added to the DOM, by using two nested `requestAnimationFrame`
calls.

This doesn't fix all of the flickering, but does alleviate a significant
amount of it regardless.

The remaining flickering shouldn't be happening, and I'm tempted to
say it's a bug in Firefox: digging in with a profiler reveals trace markers
indicating that the browser at least *thinks* it's painting the image to screen,
but the image itself never actually shows up on screen for some reason.