Fix image preloading for custom poses
`getUsedImages()` wasn't being passed a stage parameter, leading to
images being fetched for the wrong stage.

Additionally, we attach Image elements used for preloading to the page
DOM inside of an invisible div attached to the document body; this
should encourage browsers to keep images in memory, and prevent the
image elements themselves from getting GC'd.