Panty & Stocking: Episode 1
Karbol and I have been working on much-needed interactions for Panty & Stocking for a while. This is the first instalment, with room for more cases to be filled out in the future. You'll get more out of a Stocking-Panty game than a Panty-Stocking game, but both have value, and there are some really funny lines in here.

This update is a little earlier than usual, but it should help players with a recently-renewed interest in this franchise.