Shrink individual selection screen cards
Reduces the vertical size of character cards on the individual select
screen, by adjusting the positioning / cropping of character images to
be further up along the body.

The font size for name and source labels has been reduced slightly, and
the footer positioning has been adjusted so that labels don't overflow
their cards.

Finally, the size of icons on each card has been reduced slightly to
give more space for character images.

The net effect of all this is that we can now fit three rows at a time
into the select screen view without scrolling.