Adjust selection card image CSS and add image adjustment options to meta.xml
Fixes the CSS for select screen images so that images are resized
according to their natural aspect ratio, as opposed to the container
image area aspect ratio.

Additionally adjusts default image positioning to be slightly higher up.

Finally, adds XML attributes to `meta.xml` for manually adjusting
selection image positioning and scale. These options can be set as attributes
`x`, `y`, and `scale` on the `<pic>` element, and are applied directly
to the image area as a CSS transform.