Add random fill feature to the individual selection screen, i.e. selecting random opponents from those visible on that screen
I re-used the Select Opponent button for this for now in order to change the design as little as possible. As long as no character has been clicked, you can click "Random Fill (filtered)". I also made it possible to clear the detail view, so "Random Fill (filtered)" becomes available again, either by clicking the currently viewed character again or by clicking an empty part of the list, so you don't have to select the character and remove it again if you accidentally clicked it.

Whether clicking the currently viewed character a second time (without properly double-clicking) should reset the detail view instead of selecting that character can be debated.