Adjustments for select screen image display attributes
Adjusts the `scale` attribute for the metadata `<pic>` element to be
interpreted as a percentage, instead of as a 0-1 multiplier.

This is for consistency with the `<scale>` element in `meta.xml`.

Also adjusts the `x` and `y` attributes to be read as raw numbers that
are later used as percentages, instead of as raw strings for CSS; this
is to simplify CE support.