Takes Morrigan and Yang offline
This is the next round of culling. As there are three slots as usual, but now only two culling candidates remaining under this metric, obviously these would be the ones with the "lowest re-sort score". Therefore, we're removing them early as there's no point including them in the re-sort poll.

Blake (Classic) is also being "culled", but since she is already offline, all this means is that she will not come back online even if the Blake rework falls off testing.