Add automatic text hiding and dialogue bubble shrinking based on image sizes
This adds an optional new per-stage character attribute, specified in meta.xml as
`<required-image-height>`. This attribute, if present, specifies a
minimum amount of space that must be free of obstruction when displaying
dialogue in game. This attribute is specified in units of source image
pixels (prior to any rescaling).

In general, this corresponds to how tall a character is in their source images, excluding empty space.

If present, OpponentDisplay will use this value to automatically calculate
a maximum dialogue bubble height. Dialogue lines that overflow this maximum height
will be clipped to prevent the bubble from overlapping the character image.

Dialogue that is hidden in this way is masked with a gradual "fade out"
effect to indicate that it's being cut off, with an additional "More..."
label as a further hint to the user that they should look for more.

The full text is shown when hovering over the dialogue bubble (or upon
tapping the dialogue bubble, on mobile). This additionally forces the
dialogue bubble to the front of the layer stack to ensure the text is visible.

This also adds logic for automatically shrinking the minimum height for
individual dialogue bubbles, in cases where the default minimum dialogue
bubble size would overlap the character's required image height.
This occurs in tandem with automatic text hiding (if necessary).

This implementation could probably be simplified; I'll work on that later.